api/requirements.txt
tensorflow==2.9.1
PyYAML==6.0.1 
numpy==1.26.4
Pillow==9.3.0